refactor: separated auth controller from main.py to its dedicated file
